pyproject.toml
[project]
name = "pdf-translator"
version = "0.1.0"
description = "PDF translator using Gemini API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai",
    "python-dotenv",
    "pypdf2",
    "tqdm"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
exclude = []
